+++ src/app/api/practice.ts
@@ -22,3 +22,29 @@ export async function submitPracticePost({
 
   return await res.json();
 }
+
+import { createClient } from "@supabase/supabase-js";
+
+const supabase = createClient(
+  process.env.NEXT_PUBLIC_SUPABASE_URL!,
+  process.env.NEXT_PUBLIC_SUPABASE_KEY!
+);
+
+export async function fetchPracticePosts() {
+  const { data, error } = await supabase
+    .from("html_practice_posts")
+    .select("*")
+    .eq("is_public", true)
+    .order("created_at", { ascending: false });
+
+  if (error) throw new Error(error.message);
+  return data;
+}
+
+export async function likePracticePost(postId: number) {
+  const { error } = await supabase.rpc("increment_good_count", {
+    post_id: postId,
+  });
+
+  if (error) throw new Error(error.message);
+}
